feat : create update and delete feature for koor teknis

+++ routes/api.php
@@ -36,10 +36,10 @@ Route::middleware('auth:sanctum')->group(function() {
 
 Route::middleware('auth:sanctum')->group(function() {
     Route::get('koorteknis/{koorteknis_id}/document', [KoorTeknisController::class, 'index']);
-    Route::get('', [KoorTeknisController::class, 'show']);
+    // Route::get('', [KoorTeknisController::class, 'show']);
     Route::post('koorteknis/document/{document_id}', [KoorTeknisController::class, 'store']);
     Route::patch('koorteknis/{user_id}/document/{document_id}/edit', [KoorTeknisController::class, 'update']);
-    Route::delete('', [KoorTeknisController::class, 'destroy']);
+    Route::delete('koorteknis/{user_id}/document/{document_id}', [KoorTeknisController::class, 'destroy']);
 });
 
 Route::middleware('auth:sanctum')->group(function() {
